Move OfflineRequestSnapshot to its own file
